refactor: update productType filename
