+++ .github/whitelist-capital-letters.txt
@@ -38,7 +38,6 @@
 ./libs/common/src/misc/linkedFieldOption.decorator.ts
 ./libs/common/src/misc/serviceUtils.ts
 ./libs/common/src/misc/serviceUtils.spec.ts
-./libs/common/src/enums/productType.ts
 ./libs/common/src/enums/uriMatchType.ts
 ./libs/common/src/enums/secureNoteType.ts
 ./libs/common/src/enums/storageLocation.ts
