+++ apps/web/src/app/accounts/trial-initiation/trial-initiation.component.ts
@@ -14,7 +14,7 @@ import { PolicyData } from "@bitwarden/common/admin-console/models/data/policy.d
 import { MasterPasswordPolicyOptions } from "@bitwarden/common/admin-console/models/domain/master-password-policy-options";
 import { Policy } from "@bitwarden/common/admin-console/models/domain/policy";
 import { PlanType } from "@bitwarden/common/billing/enums/plan-type";
-import { ProductType } from "@bitwarden/common/enums/productType";
+import { ProductType } from "@bitwarden/common/enums/product-type";
 import { ReferenceEventRequest } from "@bitwarden/common/models/request/reference-event.request";
 
 import { RouterService } from "./../../core/router.service";
