+++ apps/web/src/app/admin-console/settings/create-organization.component.ts
@@ -3,7 +3,7 @@ import { ActivatedRoute } from "@angular/router";
 import { first } from "rxjs/operators";
 
 import { PlanType } from "@bitwarden/common/billing/enums/plan-type";
-import { ProductType } from "@bitwarden/common/enums/productType";
+import { ProductType } from "@bitwarden/common/enums/product-type";
 
 import { OrganizationPlansComponent } from "../../billing/settings/organization-plans.component";
