+++ libs/common/src/admin-console/models/data/organization.data.ts
@@ -1,4 +1,4 @@
-import { ProductType } from "../../../enums/productType";
+import { ProductType } from "../../../enums/product-type";
 import { OrganizationUserStatusType } from "../../enums/organization-user-status-type";
 import { OrganizationUserType } from "../../enums/organization-user-type";
 import { PermissionsApi } from "../api/permissions.api";
