+++ libs/common/src/admin-console/models/domain/organization.ts
@@ -1,6 +1,6 @@
 import { Jsonify } from "type-fest";
 
-import { ProductType } from "../../../enums/productType";
+import { ProductType } from "../../../enums/product-type";
 import { OrganizationUserStatusType } from "../../enums/organization-user-status-type";
 import { OrganizationUserType } from "../../enums/organization-user-type";
 import { PermissionsApi } from "../api/permissions.api";
